flask-server/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-session
redis
flask-cors
<<<<<<< Updated upstream
spotipy
=======
spotipy
openai
axios
Pillow
>>>>>>> Stashed changes
